Implementation of Cells, Pickaxe Skins and Boosters alongside bombs and minor changes to database system to allow for leaderboard implementation.

+++ src/main/java/dev/foggies/prisoncore/pickaxe/enchants/api/AbstractEnchant.java
@@ -24,7 +24,7 @@ public abstract class AbstractEnchant {
     private final long startLevel;
     private final long maxLevel;
     private final long cost;
-    private final long costMultiplier;
+    private final float costMultiplier;
 
     private final float chance;
